# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type QuoteStatus = typeof QuoteStatus[keyof typeof QuoteStatus];


export const QuoteStatus = {
  draft: 'draft',
  submitted: 'submitted',
  confirmed: 'confirmed',
  completed: 'completed',
} as const;

export interface Quote {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  projectAddress: string;
  status: QuoteStatus;
  /** @nullable */
  estimatedTotal?: number | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface QuoteItemInput {
  windowType: string;
  quantity: number;
  size?: string;
  interiorColor?: string;
  exteriorColor?: string;
  gridStyle?: string;
  glassCoating?: string;
  trimPackage?: string;
  addOns?: string;
}

export interface QuoteInput {
  fullName: string;
  email: string;
  phone: string;
  projectAddress: string;
  items: QuoteItemInput[];
}

export type QuoteUpdateStatus = typeof QuoteUpdateStatus[keyof typeof QuoteUpdateStatus];


export const QuoteUpdateStatus = {
  draft: 'draft',
  submitted: 'submitted',
  confirmed: 'confirmed',
  completed: 'completed',
} as const;

export interface QuoteUpdate {
  fullName?: string;
  email?: string;
  phone?: string;
  projectAddress?: string;
  status?: QuoteUpdateStatus;
  items?: QuoteItemInput[];
}

export interface QuoteItem {
  id: number;
  quoteId: number;
  windowType: string;
  quantity: number;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  interiorColor?: string | null;
  /** @nullable */
  exteriorColor?: string | null;
  /** @nullable */
  gridStyle?: string | null;
  /** @nullable */
  glassCoating?: string | null;
  /** @nullable */
  trimPackage?: string | null;
  /** @nullable */
  addOns?: string | null;
  /** @nullable */
  unitPrice?: number | null;
  /** @nullable */
  lineTotal?: number | null;
}

export interface QuoteSummary {
  totalQuotes: number;
  submittedQuotes: number;
  confirmedQuotes: number;
  completedQuotes: number;
  /** @nullable */
  totalEstimatedValue?: number | null;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  scheduled: 'scheduled',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface Appointment {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  address: string;
  date: string;
  timeSlot: string;
  consultationType: string;
  duration: string;
  status: AppointmentStatus;
  createdAt: string;
}

export interface AppointmentInput {
  fullName: string;
  email: string;
  phone: string;
  address: string;
  date: string;
  timeSlot: string;
  consultationType: string;
  duration?: string;
}

export interface AvailabilitySlot {
  date: string;
  timeSlot: string;
  isAvailable: boolean;
}

export type GetAvailabilityParams = {
from: string;
to: string;
};