# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from './appointmentStatus';

export interface Appointment {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  address: string;
  date: Date;
  timeSlot: string;
  consultationType: string;
  duration: string;
  status: AppointmentStatus;
  createdAt: Date;
}