# lib/api-zod/src/generated/types/appointmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentInput {
  fullName: string;
  email: string;
  phone: string;
  address: string;
  date: Date;
  timeSlot: string;
  consultationType: string;
  duration?: string;
}