# lib/api-zod/src/generated/types/error.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */

export interface Error {
  error: string;
}