# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteStatus } from './quoteStatus';

export interface Quote {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  projectAddress: string;
  status: QuoteStatus;
  /** @nullable */
  estimatedTotal?: number | null;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
}