# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteItemInput } from './quoteItemInput';

export interface QuoteInput {
  fullName: string;
  email: string;
  phone: string;
  projectAddress: string;
  items: QuoteItemInput[];
}