# lib/api-zod/src/generated/types/quoteItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteItem {
  id: number;
  quoteId: number;
  windowType: string;
  quantity: number;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  interiorColor?: string | null;
  /** @nullable */
  exteriorColor?: string | null;
  /** @nullable */
  gridStyle?: string | null;
  /** @nullable */
  glassCoating?: string | null;
  /** @nullable */
  trimPackage?: string | null;
  /** @nullable */
  addOns?: string | null;
  /** @nullable */
  unitPrice?: number | null;
  /** @nullable */
  lineTotal?: number | null;
}