# lib/api-zod/src/generated/types/quoteItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteItemInput {
  windowType: string;
  quantity: number;
  size?: string;
  interiorColor?: string;
  exteriorColor?: string;
  gridStyle?: string;
  glassCoating?: string;
  trimPackage?: string;
  addOns?: string;
}