# lib/api-zod/src/generated/types/quoteSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteSummary {
  totalQuotes: number;
  submittedQuotes: number;
  confirmedQuotes: number;
  completedQuotes: number;
  /** @nullable */
  totalEstimatedValue?: number | null;
}