# lib/api-zod/src/generated/types/quoteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for TP Windows & Doors Quote Funnel
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteItemInput } from './quoteItemInput';
import type { QuoteUpdateStatus } from './quoteUpdateStatus';

export interface QuoteUpdate {
  fullName?: string;
  email?: string;
  phone?: string;
  projectAddress?: string;
  status?: QuoteUpdateStatus;
  items?: QuoteItemInput[];
}